Add files via upload
Create http server using node and to verify open your web browser and go to 'http://localhost:8080'
The first line is just getting the http module and saving it to the variable 'http'. The http is included with Node.js to make it easy for us to create Node.js applications.
Then http server is created  by calling it's function 'createServer'. This function returns an http object and takes  function in parameter.
Next the function 'listen' on new server object is called which takes in value 8080 and tells server to which port to listen . This will connect browser to http://localhost:8080